File relocation and deactivation bug correction
