+++ includes/admin/class-deactivator.php
@@ -25,7 +25,7 @@ class Disciple_Tools_Deactivator {
 	public static function deactivate() {
 
         // Reset roles and capabilities
-        require_once('../admin/class-roles.php');
+        require_once('class-roles.php');
         $roles = Disciple_Tools_Roles::instance();
         $roles->reset_roles();
 	}
