+++ js/disciple-tools-admin.js
@@ -1,13 +1,2 @@
 
-function findGroupUsers(group_id) {
-    jQuery.post(
-        ajaxurl,
-        {
-            'action': 'list_group_users',
-            'data':   group_id
-        },
-        function(response) { alert(response)}
-
-    );
-}
